+++ CHANGELOG.md
@@ -3,8 +3,15 @@
 ## 3.10.0/4.0.0 (Upcoming)
 
 ### Major New features
+* Yang "min-element" and "max-element" feature supported
+  * According to RFC 7950 7.7.4 and 7.7.5
+  * See (tests)[test/test_minmax.sh]
+  * The following cornercases are not supported:
+    * Check for min-elements>0 for empty lists on top-level
+    * Check for min-elements>0 for empty lists in choice/case
 * Yang "unique" feature supported
-  * See RFC 7950 7.8.3
+  * According to RFC 7950 7.8.3
+  * See (tests)[test/test_unique.sh]
 * Persistent CLI history: [Preserve CLI command history across sessions. The up/down arrows](https://github.com/clicon/clixon/issues/79)
   * The design is similar to bash history:
       * The CLI loads/saves its complete history to a file on entry and exit, respectively
@@ -46,6 +53,8 @@
 	
 ### API changes on existing features (you may need to change your code)
 
+* New Clixon Yang RPC: ping. To check if backup is running.
+  * Try with `<rpc xmlns="http://clicon.org/lib"><ping/></rpc>]]>]]>`
 * Restconf with startup feature will now copy all edit changes to startup db (as it should according to RFC 8040)
 * Netconf Startup feature is no longer hardcoded, you need to explicitly enable it (See RFC 6241, Section 8.7)
   * Enable in config file with: `<CLICON_FEATURE>ietf-netconf:startup</CLICON_FEATURE>`, or use `*:*`
@@ -116,6 +125,8 @@
 
 ### Minor changes
 
+* Netconf error handling modified
+  * New option -e added. If set, the netconf client returns -1 on error.
 * A new minimal "hello world" example has been added
 * Experimental customized error output strings, see [lib/clixon/clixon_err_string.h]
 * Empty leaf values, eg <a></a> are now checked at validation.
@@ -143,6 +154,7 @@
 * Added libgen.h for baseline()
 	
 ### Corrected Bugs
+* Fixed support for multiple datanodes in a choice/case statement. Only single datanode was supported.
 * Fixed an ordering problem showing up in validate/commit callbacks. If two new items following each order (yang-wise), only the first showed up in the new-list. Thanks achernavin!
 * Fixed a problem caused by recent sorting patches that made "ordered-by user" lists fail in some cases, causing multiple list entries with same keys. NACM being one example. Thanks vratnikov!
 * [Restconf does not handle startup datastore according to the RFC](https://github.com/clicon/clixon/issues/74)